requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
distro==1.9.0
easyocr==1.7.2
et-xmlfile==1.1.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imageio==2.36.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.1
numpy==2.1.2
openai==1.52.2
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PySocks==1.7.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
scikit-image==0.24.0
scipy==1.14.1
selenium==4.25.0
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.13.1
tifffile==2024.9.20
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.0.5
wsproto==1.2.0
